requirements.txt
Django>=6.0.2
djangorestframework>=3.15.0
django-environ>=0.11.2
django-extensions>=3.2.3
celery>=5.4.0
redis>=5.0.0
psycopg2-binary>=2.9.9
channels>=4.1.0
daphne>=4.1.2
gunicorn>=22.0.0
python-json-logger>=2.0.7
channels-redis>=4.2.0
django-cors-headers>=4.3.0
django-redis>=5.4.0
